+++ api/routes/prescription.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../db');
+const auth = require('../auth');
 const jwt = require('jsonwebtoken');
 
 /**
@@ -54,8 +55,8 @@ router.get('/', (req, res, next) => {
 });
 
 /**
- * @api {get} /api/prescription/:prescriptionId Get info about prescription
- * @apiName GetPrescription
+ * @api {get} /api/prescription/:prescriptionId Get info about prescription by ID
+ * @apiName GetPrescriptionById
  * @apiGroup Prescription
  *
  * @apiParam {Number} prescriptionId Prescription ID
@@ -83,36 +84,50 @@ router.get('/', (req, res, next) => {
  * }
  */
 
-router.get('/:prescriptionId', (req, res, next) => {
+router.get('/:prescriptionId', auth({ roles: [2, 3, 4] }), (req, res, next) => {
     const prescriptionId = req.params.prescriptionId;
 
-    db.query(`SELECT * FROM prescriptions WHERE id='${prescriptionId}'`, (err, data) => {
-        if (!err) {
-            if (data.length < 1) {
-                res.status(404).json({
-                    'message': 'Prescription with provided id not found'
-                });
+    if (prescriptionId !== undefined) {
+        db.query(`SELECT * FROM prescriptions WHERE id='${prescriptionId}'`, (err, data) => {
+            if (!err) {
+                switch (data.length) {
+                    case 1:
+                        data = data[0];
+                        res.status(200).json({
+                            'id': data.id,
+                            'doctorId': data.doctorId,
+                            'patientId': data.patientId,
+                            'medicineId': data.medicineId,
+                            'description': data.description,
+                            'medicine_taking_frequency': data.medicine_taking_frequency
+                        });
+                        break;
+                    case 0:
+                        res.status(404).json({
+                            'message': 'Prescription with provided id not found'
+                        });
+                        break;
+                    default:
+                        res.status(500).json({
+                            'message': 'Database error'
+                        });
+                        break;
+                }
             } else {
-                data = data[0];
-                res.status(200).json({
-                    'id': data.id,
-                    'doctorId': data.doctorId,
-                    'patientId': data.patientId,
-                    'medicineId': data.medicineId,
-                    'description': data.description,
-                    'medicine_taking_frequency': data.medicine_taking_frequency
+                res.status(500).json({
+                    'error': err
                 });
             }
-        } else {
-            res.status(500).json({
-                'error': err
-            });
-        }
-    });
+        });
+    } else {
+        res.status(400).json({
+            'message': 'Not enough data provided'
+        });
+    }
 });
 
 /**
- * @api {get} /api/patient/:patientId Get info about prescription
+ * @api {get} /api/prescription/patient/:patientId Get info about prescription by patientId
  * @apiName GetPrescriptionByPatientId
  * @apiGroup Prescription
  *
@@ -141,26 +156,40 @@ router.get('/:prescriptionId', (req, res, next) => {
  * }
  */
 
-router.get('/patient/:patientId', (req, res, next) => {
+router.get('/patient/:patientId', auth({ roles: [2, 3, 4] }), (req, res, next) => {
     const patientId = req.params.patientId;
 
-    db.query(`SELECT * FROM prescriptions WHERE patientId='${patientId}'`, (err, data) => {
-       if (!err) {
-           if (data.length > 0) {
-               res.status(200).json({
-                   'data': data
-               });
-           } else {
-               res.status(404).json({
-                   'message': 'No prescriptions found for provided patientId'
-               });
-           }
-       } else {
-           res.status(500).json({
-               'error': err
-           })
-       }
-    });
+    if (patientId !== undefined) {
+        db.query(`SELECT * FROM prescriptions WHERE patientId='${patientId}'`, (err, data) => {
+            if (!err) {
+                switch (data.length) {
+                    case 1:
+                        res.status(200).json({
+                            'data': data
+                        });
+                        break;
+                    case 0:
+                        res.status(404).json({
+                            'message': 'No prescriptions found for provided patientId'
+                        });
+                        break;
+                    default:
+                        res.status(500).json({
+                            'message': 'Database error'
+                        });
+                        break;
+                }
+            } else {
+                res.status(500).json({
+                    'error': err
+                })
+            }
+        });
+    } else {
+        res.status(400).json({
+            'message': 'Not enough data provided'
+        });
+    }
 });
 
 /**
@@ -172,7 +201,7 @@ router.get('/patient/:patientId', (req, res, next) => {
  * @apiParam {Number} patientId Patient ID
  * @apiParam {Number} medicineId Medicine ID
  * @apiParam {String} description Description
- * @apiParam {String} medicineTakingFrequency Medicine taking frequency
+ * @apiParam {String} medicine_taking_frequency Medicine taking frequency
  *
  * @apiSuccessExample {json} Success
  * HTTP/1.1 201 OK
@@ -181,24 +210,37 @@ router.get('/patient/:patientId', (req, res, next) => {
  * }
  */
 
-router.post('/add', (req, res, next) => {
+router.post('/add', auth({ roles: [3, 4] }), (req, res, next) => {
     const doctorId = req.body.doctorId;
     const patientId = req.body.patientId;
     const medicineId = req.body.medicineId;
     const description = req.body.description;
-    const medicineTakingFrequency = req.body.medicineTakingFrequency;
+    const medicine_taking_frequency = req.body.medicine_taking_frequency;
 
-    db.query(`INSERT INTO prescriptions VALUES(NULL, '${doctorId}', '${patientId}', '${medicineId}', '${description}', '${medicineTakingFrequency}')`, (err, data) => {
-        if (!err) {
-            res.status(201).json({
-                'message': 'Prescription created successfully'
-            });
-        } else {
-            res.status(500).json({
-                'error': err
-            });
-        }
-    });
+    if (doctorId !== undefined &&
+        patientId !== undefined &&
+        medicineId !== undefined &&
+        description !== undefined &&
+        medicine_taking_frequency !== undefined
+    ) {
+        db.query(`INSERT INTO prescriptions VALUES(NULL, '${doctorId}', '${patientId}', '${medicineId}', '${description}', '${medicine_taking_frequency}')`, (err, data) => {
+            if (!err) {
+                res.status(201).json({
+                    'message': 'Prescription created successfully'
+                });
+            } else {
+                res.status(500).json({
+                    'error': err
+                });
+            }
+        });
+    } else {
+        res.status(400).json({
+            'message': 'Not enough data provided'
+        });
+    }
+
+
 });
 
 /**
@@ -215,25 +257,40 @@ router.post('/add', (req, res, next) => {
  * }
  */
 
-router.delete('/:prescriptionId', (req, res, next) => {
+router.delete('/:prescriptionId', auth({ roles: [4] }), (req, res, next) => {
     const prescriptionId = req.params.prescriptionId;
-    db.query(`DELETE FROM prescriptions WHERE id='${prescriptionId}'`, (err, data) => {
-        if (!err) {
-            if (data.length < 1) {
-                res.status(404).json({
-                    'message': 'Prescription not found'
-                });
+
+    if (prescriptionId !== undefined) {
+        db.query(`DELETE FROM prescriptions WHERE id='${prescriptionId}'`, (err, data) => {
+            if (!err) {
+                switch (data.length) {
+                    case 1:
+                        res.status(200).json({
+                            'message': 'Prescription successfully deleted'
+                        });
+                        break;
+                    case 0:
+                        res.status(404).json({
+                            'message': 'Prescription not found'
+                        });
+                        break;
+                    default:
+                        res.status(500).json({
+                            'message': 'Database error'
+                        });
+                        break;
+                }
             } else {
-                res.status(200).json({
-                    'message': 'Prescription successfully deleted'
+                res.status(500).json({
+                    'error': err
                 });
             }
-        } else {
-            res.status(500).json({
-                'error': err
-            });
-        }
-    });
+        });
+    } else {
+        res.status(400).json({
+            'message': 'Not enough data provided'
+        });
+    }
 });
 
 module.exports = router;
